Completed most taks.

+++ project/src/io.c
@@ -11,7 +11,7 @@ int printChangeMessage(const char *path, mode_t previous_mode, mode_t new_mode, 
     int n = snprintf(info, size - 1, "mode of '%s' changed from %#o (%s) to %#o (%s)\n", path,
                      previous_mode, previous_mode_str, new_mode,
                      new_mode_str);
-    if (n < 0 || n >= size - 1) return -1;
+    if (n < 0 || n >= (int) size - 1) return -1;
     return 0;
 }
 
@@ -20,7 +20,7 @@ int printRetainMessage(const char *path, mode_t mode, char *info, unsigned int s
     char mode_str[] = "---------";
     if (parseModeToString(mode, mode_str)) return -1;
     int n = snprintf(info, size - 1, "mode of '%s' retained as %#o (%s)\n", path, mode, mode_str);
-    if (n < 0 || n >= size - 1) return -1;
+    if (n < 0 || n >= (int) size - 1) return -1;
     return 0;
 }
 
@@ -33,7 +33,7 @@ int printFailedMessage(const char *path, mode_t new_mode) {
     int n = snprintf(info, sizeof(info) - strlen(info) - 1, "failed to change mode of '%s' changed to %#o (%s)\n", path,
                      new_mode, new_mode_str);
     if (n != 0) return -1;
-    if (write(STDOUT_FILENO, info, strlen(info) == -1))return -1;
+    if (write(STDOUT_FILENO, info, strlen(info) == (size_t) -1)) return -1;
     return 0;
 }
 
@@ -60,7 +60,7 @@ int printNoPermissionMessage(const char *path) {
     strncpy(info, "xmod: changing permissions of '", sizeof(info) - strlen(info) - 1);
     strncat(info, path, sizeof(info) - strlen(info) - 1);
     strncat(info, "': Operation not permitted\n", sizeof(info) - strlen(info) - 1);
-    if (write(STDERR_FILENO, info, strlen(info) == -1)) return -1;
+    if (write(STDERR_FILENO, info, strlen(info) == (size_t) -1)) return -1;
     return 0;
 }
 
@@ -95,6 +95,6 @@ int printMessage(mode_t new_mode, mode_t old_mode, const command_t *command, boo
     } else if (command->changes && new_mode != old_mode) {
         if (printChangeMessage(command->path, old_mode, new_mode, info, sizeof(info) - strlen(info) - 1)) return -1;
     }
-    if (write(STDOUT_FILENO, info, strlen(info) == -1)) return -1;
+    if (write(STDOUT_FILENO, info, strlen(info) == (size_t) -1)) return -1;
     return 0;
 }
